+++ src/js/mixin/class.js
@@ -1,7 +1,7 @@
-import { addClass, hasClass } from 'uikit-util';
+import { addClass } from 'uikit-util';
 
 export default {
     connected() {
-        !hasClass(this.$el, this.$name) && addClass(this.$el, this.$name);
+        addClass(this.$el, this.$options.id);
     },
 };